+++ libs/js-sdk/src/update-strategies/createManualUpdateStrategy.ts
@@ -1,4 +1,4 @@
-import { createEnsureSingle } from '../ensure-single'
+import { createEnsureSingleWithBackoff } from '../ensure-single'
 import { fetchFeaturesConfigurationViaHttp } from '../utils/fetchFeaturesConfiguration'
 import type { EffectiveConfigUpdateStrategy } from './update-strategies'
 import { updatesLog } from './updates-log'
@@ -10,14 +10,16 @@ export function createManualUpdateStrategy(
     httpEndpoint: string,
 ): EffectiveConfigUpdateStrategy {
     let etag: undefined | string
-    let fetchUpdatesSingle: undefined | (() => Promise<void>)
+    let fetchUpdatesSingle:
+        | undefined
+        | (() => Promise<void>)
 
     return {
         async connect(stateStore) {
             // Force update
             etag = undefined
             // Ensure that we don't trigger another request while one is in flight
-            fetchUpdatesSingle = createEnsureSingle(async () => {
+            fetchUpdatesSingle = createEnsureSingleWithBackoff(async () => {
                 etag = await fetchFeaturesConfigurationViaHttp(
                     httpEndpoint,
                     stateStore.audiences,