+++ libs/node-sdk/src/utils/fetchFeaturesConfiguration.ts
@@ -1,4 +1,7 @@
-import type { FeatureConfiguration } from '@featureboard/contracts'
+import {
+    TooManyRequestsError,
+    type FeatureConfiguration,
+} from '@featureboard/contracts'
 import type { AllFeatureStateStore } from '../feature-state-store'
 import { httpClientDebug } from './http-log'
 
@@ -8,12 +11,13 @@ export async function fetchFeaturesConfigurationViaHttp(
     stateStore: AllFeatureStateStore,
     etag: string | undefined,
     updateTrigger: string,
-) {
+): Promise<string | undefined> {
     httpClientDebug(
         'Fetching updates: trigger=%s, lastModified=%s',
         updateTrigger,
         etag,
     )
+
     const response = await fetch(allEndpoint, {
         method: 'GET',
         headers: {
@@ -21,6 +25,34 @@ export async function fetchFeaturesConfigurationViaHttp(
             ...(etag ? { 'if-none-match': etag } : {}),
         },
     })
+
+    if (response.status === 429) {
+        // Too many requests
+        const retryAfterHeader = response.headers.get('Retry-After')
+        const retryAfterInt = retryAfterHeader
+            ? parseInt(retryAfterHeader, 10)
+            : 60
+        const retryAfter =
+            retryAfterHeader && !retryAfterInt
+                ? new Date(retryAfterHeader)
+                : new Date()
+
+        if (retryAfterInt) {
+            const retryAfterTime = retryAfter.getTime() + retryAfterInt * 1000
+            retryAfter.setTime(retryAfterTime)
+        }
+
+        throw new TooManyRequestsError(
+            `Failed to get latest features: Service returned ${
+                response.status
+            }${response.statusText ? ' ' + response.statusText : ''}. ${
+                retryAfterHeader
+                    ? 'Retry after: ' + retryAfter.toUTCString()
+                    : ''
+            } `,
+            retryAfter,
+        )
+    }
 
     if (response.status !== 200 && response.status !== 304) {
         throw new Error(